:truck: Move AuthUser into models folder.
Why? all the other models are there and I want to add a "SimpleUser" as well.
