+++ api/src/data/models/index.ts
@@ -1,3 +1,4 @@
 export * from "./asset_item";
 export * from "./asset_tag_printer";
 export * from "./asset_transfer";
+export * from "./auth-user";
